Save bouts and trajectories
Save bouts in Bouts.txt (text file)
Save trajectories in Trajectory.dat (binary file)

+++ Vision.h
@@ -60,7 +60,7 @@ public:
 
     // Time
     long int frame;
-    double time;
+    qint64 time;
 
     // Calibration
     double thresholdCalibration;
